rename to FieldSchema to be consistent with proto.
